+++ packages/services/src/nodes/notebook.ts
